[WISE-245] moved WiseWebServiceException into org.jboss.wise.core.exception
