+++ src/main/java/org/jboss/wise/Wise.gwt.xml
@@ -7,15 +7,15 @@
   <!-- the theme of your GWT application by uncommenting          -->
   <!-- any one of the following lines.                            -->
   <stylesheet src='../Main.css' />
-  <!-- <inherits name='com.google.gwt.user.theme.standard.Standard'/> -->
 
   <!-- Other module inherits                                      -->
   <!-- Specify the app entry point class.  -->
   <entry-point class='org.jboss.wise.gwt.client.Main'/>
 
   <!-- Specify the paths for translatable code                    -->
-  <source path='client'/>
-  <source path='shared'/>
-  <source path='test'/>
+  <source path='gwt/client'/>
+  <source path='gwt/shared'/>
+  <source path='gwt/test'/>
+  <source path='core/exception'/>
 
 </module>
